+++ R/solve_SARSOP.R
@@ -1,0 +1,123 @@
+#' Solve a POMDP Problem using SARSOP
+#' 
+#' This function uses the C++ implementation of the SARSOP algorithm 
+#' by Kurniawati, Hsu and Lee (2008) interfaced in 
+#' package \pkg{sarsop}
+#' to solve infinite horizon problems that are formulated as partially observable Markov
+#' decision processes (POMDPs). The result is an optimal or approximately
+#' optimal policy.
+#'
+#' @param model a POMDP problem specification created with \code{\link{POMDP}}.
+#' Alternatively, a POMDP file or the URL for a POMDP file can be specified.
+#' @param method string; there is only one method available called \code{"sarsop"}.
+#' @param horizon need to be \code{Inf}.
+#' @param discount discount factor in range [0, 1]. If \code{NULL}, then the
+#' discount factor specified in \code{model} will be used.
+#' @param terminal_values needs to ne \code{NULL}. SARSOP does not use terminal values.
+#' @param digits precision used when writing POMDP files (see
+#' \code{\link{write_POMDP}}).
+#' @param parameter a list with parameters passed on to 
+#' the function \code{sarsop} in package \pkg{sarsop}.
+#' @param verbose logical, if set to \code{TRUE}, the function provides the
+#' output of the solver in the R console.
+#' @return The solver returns an object of class POMDP which is a list with the
+#' model specifications (\code{model}), the solution (\code{solution}), and the
+#' solver output (\code{solver_output}).
+#' @author Michael Hahsler
+#' @references 
+#' 
+#' Carl Boettiger, Jeroen Ooms and Milad Memarzadeh (2020). sarsop:
+#' Approximate POMDP Planning Software. R package version 0.6.6.
+#' https://CRAN.R-project.org/package=sarsop
+#' 
+#' H. Kurniawati, D. Hsu, and W.S. Lee (2008). SARSOP: Efficient point-based POMDP planning by approximating optimally reachable belief spaces. In Proc. Robotics: Science and Systems.
+#' 
+#' @examples
+#' \dontrun{
+#' # Solving the simple infinite-horizon Tiger problem
+#' data("Tiger")
+#' Tiger
+#' 
+#' sol <- solve_SARSOP(model = Tiger)
+#' sol
+#' 
+#' # look at solver output
+#' sol$solver_output
+#' 
+#' # policy (value function (alpha vectors), optimal action and observation dependent transitions)
+#' policy(sol)
+#' 
+#' # value function
+#' plot_value_function(sol, ylim = c(0,20))
+#' 
+#' # reward of the optimal policy
+#' reward(sol)
+#' }
+#' 
+#' @export
+solve_SARSOP <- function(
+  model,
+  horizon = NULL,
+  discount = NULL,
+  terminal_values = NULL,
+  method = "sarsop",
+  digits = 7,
+  parameter = NULL,
+  verbose = FALSE) {
+
+  if(!.installed("sarsop")) stop("Required package 'sarsop' is not installed.")
+  
+  if(method != "sarsop") stop("Only available method: 'sarsop'")
+  
+  # check parameters
+  if(!is.null(terminal_values)) stop("the SARSOP solver does not support terminal values.")
+  
+  if(!is.null(horizon)) model$model$horizon <- horizon
+  if(!is.infinite(model$model$horizon)) stop("the SARSOP solver only supports infinite time horizon problems.")
+  
+  if(!is.null(discount)) model$model$discount <- discount
+  
+  if(.timedependent_POMDP(model)) stop("the SARSOP solver does not support time dependent models.")
+  
+  # prepare temp files
+  tmpf <- tempfile()
+  model_file <- paste0(tmpf, '.pomdp')
+  policy_file <- paste0(tmpf, '.policyx')
+  log_file <- paste0(tmpf, '.log')
+  
+  # call SARSOP
+  write_POMDP(model, file = model_file, digits = digits)
+  res <- do.call(sarsop::pomdpsol, c(
+    list(model = model_file, output = policy_file, 
+    stdout = log_file, stderr = log_file), parameter))
+  
+  model$solver_output = readLines(log_file)
+  class(model$solver_output) <- "text"
+  if(verbose) print(model$solver_output, "\n")
+  
+  # package solution
+  policy <- sarsop::read_policyx(policy_file)
+  pg <- data.frame(node = 1:length(policy$action), 
+    action = model$model$actions[policy$action])
+  alpha <- t(policy$vectors)
+  colnames(alpha) = model$model$states
+  
+  model$solution <- structure(list(
+    method = "sarsop", 
+    parameter = parameter,
+    horizon = model$model$horizon,
+    discount = model$model$discount,
+    converged = length(grep("precision reached", res$end_condition)) == 1,
+    total_expected_reward = NA,
+    initial_belief = NA,
+    initial_pg_node = NA,
+    #  terminal_values = if(!is.null(terminal_values)) terminal_values else 0, 
+    #belief_states = belief, 
+    pg = list(pg),
+    alpha = list(alpha)
+  ), class = "POMDP_solution")
+  
+  model$solution$total_expected_reward = reward(model, model$model$start)$reward
+  
+  model
+}